oldfs: Clean embox.fs.syslib.kfile
